+++ app/(main)/(routes)/meself/chat/[directChatId]/page.tsx
@@ -58,6 +58,7 @@ const DirectChatMemberIdPage = async ({
             <DirectChannelHeader
               imageUrl={otherMember.imageUrl}
               name={otherMember.name}
+              directChannelHeaderIdProp={direct.id}
               userAvatarProp={profile.imageUrl}
               userNameProp={profile.name}
               userStatusProp={profile.status}