+++ components/chat-video-button.tsx
@@ -1,38 +1,48 @@
 "use client";
 
-import qs from "query-string";
-import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import { useSearchParams } from "next/navigation";
 import { Video, VideoOff } from "lucide-react";
 
 
 import { ActionTooltip } from "./uihelper/action-tooltip";
-
-export const ChatVideoButton = () => {
-  const pathname = usePathname();
-  const router = useRouter();
+import { useSocket } from "./providers/socket-provider";
+import { Sheet, SheetContent, SheetTrigger } from "./ui/sheet";
+import { Button } from "./ui/button";
+interface ChatVideoButtonInterface {
+  directChannelId: string;
+}
+export const ChatVideoButton = (
+  {
+    directChannelId
+  } : ChatVideoButtonInterface
+) => {
+  const { socketActual } = useSocket();
   const searchParams = useSearchParams();
 
   const isVideo = searchParams?.get("video");
+  if (!socketActual) {
+    console.log("Socket is not running")
+    return;
+  }
+
 
   const onClick = () => {
-    const url = qs.stringifyUrl({
-      url: pathname || "",
-      query: {
-        video: isVideo ? undefined : true,
-      }
-    }, { skipNull: true });
-
-    router.push(url);
+    socketActual.emit("overlay");
   }
 
   const Icon = isVideo ? VideoOff : Video;
-  const tooltipLabel = isVideo ? "End video call" : "Start video call";
 
   return (
-    <ActionTooltip side="bottom" label={tooltipLabel}>
-      <button onClick={onClick} className="hover:opacity-75 transition mr-4">
-        <Icon className="h-6 w-6 text-zinc-500 dark:text-zinc-400" />
-      </button>
-    </ActionTooltip>
+      <Sheet>
+      <SheetTrigger asChild>
+        <Button 
+        variant="ghost" size="icon" className="ml-auto">
+            <Icon className="h-6 w-6 text-zinc-500 dark:text-zinc-400" />      
+        </Button>
+      </SheetTrigger>
+      <SheetContent side="right" className="p-0 flex gap-0 ">
+
+      </SheetContent>
+    </Sheet>
   )
 }