+++ components/display/friend-profile.tsx
@@ -107,8 +107,7 @@ const FriendProfileComponent = (
                         </DropdownMenuContent>
                       </DropdownMenu>
                     </div>
-                    )}
-                    
+                    )}                    
                 </div>
                 
               )